pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "balance-sheet-playground"
version = "0.1.0"
description = "Toy balance-sheet playground for trading book scenarios"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["PyYAML>=6.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
